Created 'EnvironmentMap' class.
Allows easy creation of environment maps (from files or from images), and easy conversion to and from a variety of formats. Essentially a class wrapper on top of all the conversion functions in the repo.
Particularly convenient because it keeps track of its own format, and should be easily extended to accept additional formats.